Vmd関連のリネームとファイル分け
