+++ MikuMikuMethodsCore/VMD/IVmdFrame.cs
@@ -8,45 +8,14 @@ using System.Threading.Tasks;
 namespace MikuMikuMethods.VMD
 {
     /// <summary>
-    /// フレームの種類
-    /// </summary>
-    public enum FrameType
-    {
-        /// <summary>
-        /// カメラ
-        /// </summary>
-        Camera,
-        /// <summary>
-        /// 照明
-        /// </summary>
-        Light,
-        /// <summary>
-        /// 影
-        /// </summary>
-        Shadow,
-        /// <summary>
-        /// 表示・IK
-        /// </summary>
-        Property,
-        /// <summary>
-        /// モーフ
-        /// </summary>
-        Morph,
-        /// <summary>
-        /// モーション
-        /// </summary>
-        Motion,
-    }
-
-    /// <summary>
     /// フレームのインターフェイス
     /// </summary>
-    public interface IVocaloidFrame:IComparable<IVocaloidFrame>, ICloneable
+    public interface IVmdFrame:IComparable<IVmdFrame>, ICloneable
     {
         /// <summary>
         /// フレームの種類
         /// </summary>
-        FrameType FrameType { get; }
+        VmdFrameType FrameType { get; }
         /// <summary>
         /// フレームの名前
         /// </summary>
