+++ MikuMikuMethodsCore/VMD/VmdFrameType.cs
@@ -1,0 +1,33 @@
+﻿namespace MikuMikuMethods.VMD
+{
+    /// <summary>
+    /// フレームの種類
+    /// </summary>
+    public enum VmdFrameType
+    {
+        /// <summary>
+        /// カメラ
+        /// </summary>
+        Camera,
+        /// <summary>
+        /// 照明
+        /// </summary>
+        Light,
+        /// <summary>
+        /// 影
+        /// </summary>
+        Shadow,
+        /// <summary>
+        /// 表示・IK
+        /// </summary>
+        Property,
+        /// <summary>
+        /// モーフ
+        /// </summary>
+        Morph,
+        /// <summary>
+        /// モーション
+        /// </summary>
+        Motion,
+    }
+}
