+++ MikuMikuMethodsCore/VMD/VmdType.cs
@@ -1,0 +1,17 @@
+﻿namespace MikuMikuMethods.VMD
+{
+    /// <summary>
+    /// VMDの種類
+    /// </summary>
+    public enum VMDType
+    {
+        /// <summary>
+        /// カメラ系
+        /// </summary>
+        Camera,
+        /// <summary>
+        /// モデル系
+        /// </summary>
+        Model
+    }
+}
